requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
django-auth-ldap==4.3.0
django-crispy-forms==2.0
django-oauth2-authcodeflow==1.0.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
ecdsa==0.18.0
Flask==2.3.2
Flask-RESTful==0.3.9
flower==1.2.0
h2==4.1.0
hpack==4.0.0
humanize==4.6.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
oauthlib==3.2.2
ocpp==0.18.0
openapi-codec==1.3.2
Pillow==9.5.0
priority==1.3.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jose==3.3.0
python-ldap==3.4.3
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tornado==6.3.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
websockets==11.0
Werkzeug==2.3.4
zope.interface==6.0
